+++ src/main/java/org/plumelib/javadoc/RequireJavadoc.java
@@ -83,6 +83,14 @@ public class RequireJavadoc {
   public boolean dont_require_private;
 
   /**
+   * If true, don't check constructors with zero formal parameters. These are sometimes called
+   * "default constructors", though officially that term only refers to ones that the compiler
+   * synthesizes when the programmer didn't write one.
+   */
+  @Option("Don't report problems in constructors with zero formal parameters")
+  public boolean dont_require_noarg_constructor;
+
+  /**
    * If true, don't check trivial getters and setters.
    *
    * <p>Trivial getters and setters are of the form:
@@ -595,6 +603,9 @@ public class RequireJavadoc {
     @Override
     public void visit(ConstructorDeclaration cd, Void ignore) {
       if (dont_require_private && cd.isPrivate()) {
+        return;
+      }
+      if (dont_require_noarg_constructor && cd.getParameters().size() == 0) {
         return;
       }
       String name = cd.getNameAsString();